Refactor: 폴더구조 변경 및 네이밍 통일(#94)
